Move commitment calculations to transaction.go
